+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.2.0] - 2026-07-29
+
+### Added
+- Added `SceneVisibilityLocker` for persistent Scene view visibility and picking controls in the Unity Editor.
+- Added `WindowBoxCamera` for maintaining a target aspect ratio across editor, display, and render texture outputs.
 ## [5.1.0] - 2026-07-28
 
 ### Added